Converted java classes to kotlin, many changes

+++ app/src/main/java/com/alterneo/alterneo_app/responses/CompanyResponse.kt
@@ -17,8 +17,8 @@ class CompanyResponse(
     var company_create_date: String,
     var company_id_code: String,
     var registration_date: String,
-    var latitude: Float,
-    var longitude: Float
+    var latitude: Double,
+    var longitude: Double
 ) : BasicResponse(id) {
 
     fun toCompany(): Company {
